Update patch set 1
Patch Set 1:

Yes it does but in the 2.6.35 kernel it uses the same name as the Blade driver so I just took the Blade libsensors and removed the proximity sensor and it worked.

Renaming the driver to from "light" to "alsprox" and change it to use ABS_X instead of ABS_MISC in the kernel would allow the old sensors.default.so to work but now the g-sensor also works much better now with the new lib.

Patch-set: 1