Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I've noticed graphical artifacts in some apps when this is turned on. For example if you read a book in Aldiko Book Reader and press the Menu key. Turning it off in CMParts makes the artifacts go away.

My HTC Desire shows exactly the same artifacts (lots of white stuff that should not be there) if I enable this on it so the problem seems to be generic.

Not a showstopper but I think we should leave it off as default.

Patch-set: 1