Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Some devices got stuck on reporting 100% when we went from 2.6.32 kernel. This kernel has some new TI GAUGE code in msm_battery disabled and the TI_GAUGE driver has been changed back to how it looks in 2.6.32.

It has been confirmed that this fixes the stuck-on-100percent-problem and it does not seem to cause any ill effects on devices that didn't experience in the issue.

Basicly it brings battery reporting back to how it worked in the 2.6.32 kernel for the ZTE V9.

Patch-set: 1