Update patch set 1
Patch Set 1:

I've had the 0% a couple of times with the 2.6.35 kernel. Usually resolved by powering off and removing the battery. Others have reported this too and I think this should also fix that issue.

A handful of users got the stuck on 100% instead and two of them have allready confirmed this finally fixed their problem.

I suspect that the changes to msm_battery.c and ti_gauge.c in the 2.6.35 expects other ti_gauge.bqfs and ti_gauge.dffs files that we do not have (yet).

Patch-set: 1