Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I agree with Tom that the issue with the original patch (that also exists on my HTC Desire and propably a lot of other devices) should be fixed.

I also agree with Steve that since the original patch enables something that give graphical errors in some apps it should not be enabled by default.

I also doubt this issue will be fixed before the final 7.2 so it's propably best to abandon this patch.

The no longer used TARGET_USES_16BPPSURFACE_FOR_OPAQUE in BoardConfig I will remove in another patch that I'm working on right now.

Patch-set: 2