Update patch set 1
Patch Set 1:

Both blade and v9 exist in the CM9 tree.
https://github.com/CyanogenMod/android_device_zte_blade/tree/ics
https://github.com/CyanogenMod/android_device_zte_v9/tree/ics

Devices are no longer added to the main manifest to reduce the size of everyones local repo. 'brunch <device_name>' will pull it down for you and add it to your local manifest (eg. brunch v9).

A decision was made not to add devices to the brunch menu until the device is fully compliant with the CDD, so although these devices are in the CM9 tree, they are not displayed in the brunch list.

Patch-set: 1